#20 split mesh into base mesh and finite volumes
